Fix unit and integration tests

+++ tests/sync/connectblocks.go
@@ -107,7 +107,7 @@ func verifyBlockInfo(t *testing.T, d *db.RocksDB, h *TestHandler, rng Range) {
 func verifyTransactions(t *testing.T, d *db.RocksDB, h *TestHandler, rng Range) {
 	type txInfo struct {
 		txid     string
-		vout     uint32
+		vout     int32
 		isOutput bool
 	}
 	addr2txs := make(map[string][]txInfo)
@@ -122,13 +122,13 @@ func verifyTransactions(t *testing.T, d *db.RocksDB, h *TestHandler, rng Range) 
 		for _, tx := range block.TxDetails {
 			for _, vin := range tx.Vin {
 				for _, a := range vin.Addresses {
-					addr2txs[a] = append(addr2txs[a], txInfo{tx.Txid, vin.Vout, false})
+					addr2txs[a] = append(addr2txs[a], txInfo{tx.Txid, int32(vin.Vout), false})
 					checkMap[a] = append(checkMap[a], false)
 				}
 			}
 			for _, vout := range tx.Vout {
 				for _, a := range vout.ScriptPubKey.Addresses {
-					addr2txs[a] = append(addr2txs[a], txInfo{tx.Txid, vout.N, true})
+					addr2txs[a] = append(addr2txs[a], txInfo{tx.Txid, int32(vout.N), true})
 					checkMap[a] = append(checkMap[a], false)
 				}
 			}
@@ -136,7 +136,7 @@ func verifyTransactions(t *testing.T, d *db.RocksDB, h *TestHandler, rng Range) 
 	}
 
 	for addr, txs := range addr2txs {
-		err := d.GetTransactions(addr, rng.Lower, rng.Upper, func(txid string, vout uint32, isOutput bool) error {
+		err := d.GetTransactions(addr, rng.Lower, rng.Upper, func(txid string, vout int32, isOutput bool) error {
 			for i, tx := range txs {
 				if txid == tx.txid && vout == tx.vout && isOutput == tx.isOutput {
 					checkMap[addr][i] = true
